apps/backend/api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-jose[cryptography]==3.3.0
httpx==0.26.0
google-cloud-storage==2.14.0
google-cloud-secret-manager==2.18.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
google-auth==2.27.0
python-multipart==0.0.6
